UAS-103 Fixed details on first get method for subgroups

+++ src/main/kotlin/ucb/accounting/backend/dto/AccoSubGroupDto.kt
@@ -1,7 +1,7 @@
 package ucb.accounting.backend.dto
 
 data class AccoSubGroupDto (
-    val accountGroupId: Int?,
+    val accountGroupId: Long?,
     val accountSubGroupCode: Int?,
     val accountSubGroupName: String?
 )
